moved midi stuff to separate file and midi event polling is now on a different thread. \also decided to add debug messages for tracking dropped voices for some reason
